Update version of Helm.

+++ helm/helm.bash
@@ -70,24 +70,30 @@ helm repo list && helm repo update || true
 # if 'TILLERLESS=true' is set, run a local tiller server with the secret backend
 # see also https://github.com/helm/helm/blob/master/docs/securing_installation.md#running-tiller-locally
 if [ "$TILLERLESS" = true ]; then
-  # create tiller-namespace if it doesn't exist (helm --init would usually do this with server-side tiller'
-  if [[ -n $TILLER_NAMESPACE ]]; then
-    echo "Creating tiller namespace $TILLER_NAMESPACE"
-    kubectl get namespace $TILLER_NAMESPACE || kubectl create namespace $TILLER_NAMESPACE
-  fi
+  if [[ $HELM_VERSION =~ ^v2 ]]; then
 
-  echo "Starting local tiller server"
-  #default inherits --listen localhost:44134 and TILLER_NAMESPACE
-  #use the secret driver by default
-  tiller --storage=secret &
-  export HELM_HOST=localhost:44134
-  if [ "$DEBUG" = true ]; then
-      echo "Running: helm $@"
+    # create tiller-namespace if it doesn't exist (helm --init would usually do this with server-side tiller'
+    if [[ -n $TILLER_NAMESPACE ]]; then
+      echo "Creating tiller namespace $TILLER_NAMESPACE"
+      kubectl get namespace $TILLER_NAMESPACE || kubectl create namespace $TILLER_NAMESPACE
+    fi
+
+    echo "Starting local tiller server"
+    #default inherits --listen localhost:44134 and TILLER_NAMESPACE
+    #use the secret driver by default
+    tiller --storage=secret &
+    export HELM_HOST=localhost:44134
+    if [ "$DEBUG" = true ]; then
+        echo "Running: helm $@"
+    fi
+    helm "$@" && exitCode=$? || exitCode=$?
+    echo "Stopping local tiller server"
+    pkill tiller
+    exit $exitCode
+  else
+    helm "$@" && exitCode=$? || exitCode=$?
+    exit $exitCode
   fi
-  helm "$@" && exitCode=$? || exitCode=$?
-  echo "Stopping local tiller server"
-  pkill tiller
-  exit $exitCode
 else
   if [ "$DEBUG" = true ]; then
       echo "Running: helm $@"
